isoping: an isochronous ping-like program.
Like ping, but sends packets isochronously (equally spaced in time) in each
direction.  By being clever, we can use the known timing of each packet to
determine, on a noisy network, which direction is dropping or delaying
packets and by how much.  It should be especially useful for testing wifi.

Also unlike ping, this requires a server (ie. another copy of this
program) to be running on the remote end.

Change-Id: I5437a48603ca167a2e4f606f8d00e414753d1e71