Add a second, shorter 'sufficient' time limit.
The sufficient time limit causes isostream to bail out after the actual
test has run for a certain number of seconds. It allows multiple
isostream commands to be started in parallel (either on the same or
different devices) and run against a server limited in the number of
connections it accepts.

Example use:

    $ ./isostream -b 1 -s 60 -t 720 192.168.1.1

Tested:

   Manually run in parallel six ways (using tmux) against a server with
   a connection limit of 1 connection, and confirmed commands run to
   completion one after the other.

   This test includes lossy networks, via github.com/tylertreat/comcast

Change-Id: I26fc6fc1b8d4445d32c876d271a4d59bf30557cc