Add utility program to help run fuzzers against isoping.
host-isoping_fuzz reads packets from standard input, then sends the data
to the isoping server over a socket.  This lets fuzzers like afl easily
pass in modified data, while covering as much of the real code as
possible.  It also overrides the non-deterministic parts of the server
such as periodic secret rotation and cookie validation so as to not
confuse the fuzzers.

Change-Id: I991add34d65e159036e7279932dd15c6c68a200d